Fix parse errors under several circumstances
This change updates the parser to be able to successfully
parse .env files that have mixed double and single quotes.

Before that an .env file starting with a key value pair
where the value starts with single quotes would greedily
parse until the next occurence of another type of quotes.

The regex for those values was too greedy.
In the course of this fix we also identifie problems
with parsing of empty values. As a reaction to that
we switched to line based parsing which simplifies the
regex and makes the parsing process more tangible.